requirements.txt
python-telegram-bot==21.6
google-generativeai==0.8.3
httpx==0.27.0
